+++ appEditors/AppGerberEditor.py
@@ -3570,7 +3570,8 @@ class AppGerberEditor(QtCore.QObject):
         self.app.ui.grb_editor_menu.setDisabled(False)
         self.app.ui.grb_editor_menu.menuAction().setVisible(True)
 
-        self.app.ui.update_obj_btn.setEnabled(True)
+        self.app.ui.editor_exit_btn_ret_action.setVisible(True)
+        self.app.ui.editor_start_btn.setVisible(False)
         self.app.ui.grb_editor_cmenu.setEnabled(True)
 
         self.app.ui.grb_edit_toolbar.setDisabled(False)
@@ -3623,7 +3624,8 @@ class AppGerberEditor(QtCore.QObject):
         self.app.ui.grb_editor_menu.setDisabled(True)
         self.app.ui.grb_editor_menu.menuAction().setVisible(False)
 
-        self.app.ui.update_obj_btn.setEnabled(False)
+        self.app.ui.editor_exit_btn_ret_action.setVisible(False)
+        self.app.ui.editor_start_btn.setVisible(True)
 
         # adjust the visibility of some of the canvas context menu
         self.app.ui.popmenu_edit.setVisible(True)